+++ pinot-core/src/main/java/org/apache/pinot/core/geospatial/transform/function/StPolygonFunction.java
@@ -60,8 +60,7 @@ public class StPolygonFunction extends ConstructFromTextFunction {
         Preconditions.checkArgument(geometry instanceof Polygon, "The geometry object must be polygon");
         _results[i] = GeometrySerializer.serialize(geometry);
       } catch (ParseException e) {
-        Utils.rethrowException(
-            new RuntimeException(String.format("Failed to parse geometry from string: %s", argumentValues[i])));
+        new RuntimeException(String.format("Failed to parse geometry from string: %s", argumentValues[i]));
       }
     }
     return _results;
